refactor(network): move Throw to Internal & cleanup
